abstract out createTextBoxControl
